Phase 2a-4: Extract stats subcommand into commands/stats/ package
Split run_stats_command (~1010 lines) out of commands/__init__.py into
a focused commands/stats/ package:

  commands/stats/__init__.py    - dispatcher (run_stats_command)
  commands/stats/export.py      - was _stats_export
  commands/stats/analyze.py     - was _stats_analyze
  commands/stats/optimize.py    - was _stats_optimize
  commands/stats/visualize.py   - was _stats_visualize
  commands/stats/list.py        - was _stats_list

run_stats_command is re-exported from commands/__init__.py for
backwards compatibility. With this split commands/__init__.py is now
a thin facade (32 lines) re-exporting all five subcommand handlers.
Phase 2a is complete.

No behavior change — all 637 unit tests pass.

commands/__init__.py: 1058 -> 32 lines (-1026 over the full Phase 2a).
